renamed layer to node. still needs to be updated in some comments
