+++ source/node/algorithm/lerp.hpp
