+++ source/node/computed.hpp
