+++ source/node/constant.hpp
