+++ source/particle/water.hpp
@@ -5,13 +5,13 @@
 #include <soillib/particle/particle.hpp>
 #include <soillib/particle/cascade.hpp>
 
-#include <soillib/layer/layer.hpp>
+#include <soillib/core/node.hpp>
 
-#include <soillib/layer/cached.hpp>
-#include <soillib/layer/constant.hpp>
-#include <soillib/layer/computed.hpp>
+#include <soillib/node/cached.hpp>
+#include <soillib/node/constant.hpp>
+#include <soillib/node/computed.hpp>
 
-#include <soillib/layer/algorithm/normal.hpp>
+#include <soillib/node/algorithm/normal.hpp>
 
 #include <soillib/core/buffer.hpp>
 #include <soillib/matrix/matrix.hpp>
